requirements.txt
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==2.2.17
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
urllib3==1.26.2
django-debug-toolbar==3.2
django-debug-toolbar-template-profiler==2.0.2
wheel==0.36.2
